Separación en módulos, mejoras en generador web
