+++ tools/get_offer.py
@@ -1,14 +1,16 @@
+#!/bin/python3
+# coding: utf-8
+#
 # Uso: python3 get_offer.py *clave_de_materia1_* *clave_de_materia_2* ...
 # Todos los demás datos vienen hardcodeados
 
 # pip3 install requests beautifulsoup4
 import argparse
-import os
-import re
+import Horario
+import Materia
+from gen_cat import generarCatalogoMaterias
 import requests
 import bs4
-import json
-from datetime import time
 
 # Datos necesarios para obtener la oferta
 SERVIDOR_SIIAU = 'http://consulta.siiau.udg.mx'
@@ -18,451 +20,6 @@ CENTRO_DEF = "D"  # CUCEI
 PLAN_DEF = "ICOM"  # Computación, técnicamente no es necesario pero de todos modos lo agrego
 # Número de clases a mostrar, ponemos un valor exagerado para asegurarnos que nos muestre todos
 CANT_RESULT_DEF = 10000  # Índice clases
-# Ruta al catálogo de materias que contiene los links
-RUTA_CATALOGO_MATERIAS = os.path.join(
-    os.path.dirname(os.path.dirname(__file__)), "materias.json")
-
-
-class Horario:
-    horarios = list()
-    centro = str()
-    sesion = str()
-    horas = str()
-    dias = str()
-    edificio = str()
-    aula = str()
-    periodo = str()
-
-    def __init__(self, centro: str, sesion: str, horas: str, dias: str, edificio: str, aula: str, periodo: str):
-        self.centro = centro
-        self.sesion = sesion
-        self.horas = horas
-        self.dias = dias
-        self.edificio = self._limpiarEdificio(edificio, centro)
-        self.aula = aula
-        self.periodo = periodo
-
-    def _limpiarEdificio(self, edificio: str, centro: str):
-        if "Edificio" in edificio or "Virtual" in edificio:
-            if "Virtual" in edificio:
-                edificio = edificio.split("Virtual")[-1].strip()
-                edificio = "{}ESV{}".format(centro, edificio)
-            else:
-                edificio = edificio.split("Edificio")[-1].strip()
-                edificio = "{}ED{}".format(centro, edificio)
-        return edificio
-
-    def _procesarAmPm(self, hora: int, meridiano: str):
-        meridiano = meridiano.lower()
-        if hora == 12:
-            hora = 0 if "am" in meridiano else hora
-        elif "pm" in meridiano:
-            hora += 12
-        return hora
-
-    def obtenerSesion(self):
-        sesion = self.sesion
-        sesion = sesion if isinstance(sesion, int) else int(sesion[1:])
-        return sesion
-
-    def obtenerHoras(self):
-        horas = self.horas.lower()
-        textoHoras = horas
-        patronHoras = r"(?i)"
-        patronHoras += r"^(\d{2}):?(\d{2})\s?(?:(am|pm)?\s?)"
-        patronHoras += r"-"
-        patronHoras += r"\s?(\d{2}):?(\d{2})\s?(?:(am|pm)?\s?)"
-        coincidencia = re.match(patronHoras, horas)
-
-        if coincidencia:
-            coincHoras = [
-                int(x) if str(x).isdigit() else x
-                for x in coincidencia.groups()
-            ]
-            if coincHoras[2]:
-                coincHoras[0] = self._procesarAmPm(coincHoras[0],
-                                                   coincHoras[2])
-            if coincHoras[5]:
-                coincHoras[3] = self._procesarAmPm(coincHoras[3],
-                                                   coincHoras[5])
-            horasIni = time(coincHoras[0], coincHoras[1])
-            horasFin = time(coincHoras[3], coincHoras[4])
-            textoHoras = f'{horasIni.strftime("%I:%M %p")} - {horasFin.strftime("%I:%M %p")}'
-
-        return textoHoras
-
-    def obtenerDias(self):
-        dias = []
-        diasTexto = self.dias
-        if "." in diasTexto:
-            diasClave = {
-                "L": "Lunes",
-                "M": "Martes",
-                "I": "Miércoles",
-                "J": "Jueves",
-                "V": "Viernes",
-                "S": "Sábado",
-            }
-            diasSIIAU = re.sub(r"[\.\s]+", "", diasTexto)
-            for letra in diasSIIAU:
-                if letra in diasClave:
-                    dias.append(diasClave[letra])
-            diasTexto = ", ".join(dias)
-
-        return diasTexto
-
-    def obtenerEdificio(self):
-        edificio = self.edificio
-        centro = self.centro
-        if not ("Edificio" in edificio or "Virtual" in edificio):
-            coincVirt = re.match(fr"^{centro}ESV|^VIRTU\w*", edificio)
-            if coincVirt is not None:
-                edificio = re.sub(fr"^{centro}ESV|^VIRTU\w*", "", edificio)
-                edificio = f"Virtual {edificio}"
-            else:
-                edificio = re.sub(fr"^{centro}(ED)?", "", edificio)
-                edificio = f"Edificio {edificio}"
-        return edificio
-
-    def obtenerAula(self):
-        aula = self.aula
-        textoAula = aula
-        partes_aula = re.match("([A-Za-z]+)(\d+)", aula)
-        if partes_aula is not None:
-            alfa = partes_aula[1]
-            num = int(partes_aula[2])
-            textoAula = f"{alfa}{num:02d}"
-
-        return textoAula
-
-    def obtenerPeriodo(self):
-        return self.periodo
-
-
-class Materia:
-    centro = str()
-    nrc = str()
-    clave = str()
-    nombre = str()
-    seccion = str()
-    creditos = int()
-    cupo = int()
-    disponible = int()
-    horarios = []
-    profesor = str()
-    url = str()
-    eliminada = False
-
-    def __init__(self, centro: str, nrc: int, clave: str, nombre: str,
-                 seccion: str, creditos: int, cupo: int, disponible: int,
-                 horarios: list, profesor: str = "", eliminada: bool = False,
-                 url: str = ""):
-        self.centro = centro
-        self.nrc = nrc
-        self.clave = clave
-        self.nombre = nombre.upper()
-        self.seccion = seccion
-        self.creditos = creditos
-        self.cupo = cupo
-        self.disponible = disponible
-        self.horarios = horarios
-        self.profesor = profesor.upper()
-        self.url = url
-        self.eliminada = eliminada
-
-    def indiceValido(self, i, lista):
-        return (i is not None and 0 <= i < len(lista))
-
-    def obtenerSesion(self, indiceHorario=None, legible_por_humanos=True):
-        sesion = []
-        if (self.indiceValido(indiceHorario, self.horarios)):
-            horario = self.horarios[indiceHorario]
-            sesion = horario.obtenerSesion()
-        else:
-            for horario in self.horarios:
-                if (legible_por_humanos):
-                    sesion.append(horario.obtenerSesion())
-                else:
-                    sesion.append(horario.sesion)
-
-        return sesion
-
-    def obtenerHoras(self, indiceHorario=None, legible_por_humanos=True):
-        horas = []
-        if (self.indiceValido(indiceHorario, self.horarios)):
-            horario = self.horarios[indiceHorario]
-            horas = horario.obtenerHoras()
-        else:
-            for horario in self.horarios:
-                if (legible_por_humanos):
-                    horas.append(horario.obtenerHoras())
-                else:
-                    horas.append(horario.horas)
-
-        return horas
-
-    def obtenerDias(self, indiceHorario=None, legible_por_humanos=True):
-        dias = []
-        if self.indiceValido(indiceHorario, self.horarios):
-            horario = self.horarios[indiceHorario]
-            dias = horario.obtenerDias()
-        else:
-            for horario in self.horarios:
-                if legible_por_humanos:
-                    dias.append(horario.obtenerDias())
-                else:
-                    dias.append(horario.dias)
-        return dias
-
-    def obtenerEdificio(self, indiceHorario=None, legible_por_humanos=True):
-        edificios = []
-        if (self.indiceValido(indiceHorario, self.horarios)):
-            horario = self.horarios[indiceHorario]
-            edificios = horario.obtenerEdificio()
-        else:
-            for horario in self.horarios:
-                if legible_por_humanos:
-                    edificios.append(horario.obtenerEdificio())
-
-                else:
-                    edificios.append(horario.edificio)
-
-        return edificios
-
-    def obtenerAula(self, indiceHorario=None, legible_por_humanos=True):
-        aulas = []
-        if self.indiceValido(indiceHorario, self.horarios):
-            horario = self.horarios[indiceHorario]
-            aulas = horario.obtenerAula()
-        else:
-            for horario in self.horarios:
-                if legible_por_humanos:
-                    aulas.append(horario.obtenerAula())
-                else:
-                    aulas.append(horario.aula)
-        return aulas
-
-    def obtenerPeriodo(self, indiceHorario=None, legible_por_humanos=True):
-        periodos = []
-        if self.indiceValido(indiceHorario, self.horarios):
-            horario = self.horarios[indiceHorario]
-            periodos = horario.obtenerPeriodo()
-        else:
-            for horario in self.horarios:
-                if legible_por_humanos:
-                    periodos.append(horario.obtenerPeriodo())
-                else:
-                    periodos.append(horario.periodo)
-        return periodos
-
-    def obtenerNombre(self):
-        return self.nombre.title()
-
-    def obtenerProfesor(self):
-        return self.profesor.title()
-
-    def obtenerVersionHumana(self):
-        """Devuelve un diccionario con los detalles de la materia en formato legible por humanos"""
-        mat_legible_por_humanos = {
-            "centro": self.centro,
-            "seccion": self.seccion,
-            "materia": self.obtenerNombre(),
-            "cupo": self.cupo,
-            "disponible": self.disponible,
-            "creditos": self.creditos,
-            "nrc": self.nrc,
-            "clave": self.clave,
-            "sesion": self.obtenerSesion(),
-            "horas": self.obtenerHoras(),
-            "dias": self.obtenerDias(),
-            "edificio": self.obtenerEdificio(),
-            "aula": self.obtenerAula(),
-            "periodo": self.obtenerPeriodo(),
-            "profesor": self.obtenerProfesor(),
-            "url": self.url,
-            "eliminada": self.eliminada,
-        }
-        return mat_legible_por_humanos
-
-
-def cargarMaterias(clasificarMaterias: bool = True, objetosMateria: bool = False):
-    materias_cargadas = {}
-    try:
-        if os.path.exists(RUTA_CATALOGO_MATERIAS):
-            with open(RUTA_CATALOGO_MATERIAS, "r") as f:
-                materias_cargadas = json.loads(f.read())
-            if objetosMateria or not clasificarMaterias:
-                materias_cargadas = aplanarDictMaterias(materias_cargadas)
-            if objetosMateria:
-                materias_cargadas = dictAMaterias(
-                    materias_cargadas, clasificarMaterias)
-
-    except Exception:
-        import traceback
-        print("Error: No se pudieron cargar las materias del JSON")
-        print("Detalles:")
-        print(traceback.format_exc())
-
-    return materias_cargadas
-
-
-def guardarMaterias(materiasAGuardar: dict | list):
-    """
-    Guarda una lista de objetos materia, una lista de dicts
-    de materias, o un dict con la estructura de catálogo
-    en el JSON de materias
-    """
-    correcto = True
-    try:
-        if isinstance(materiasAGuardar, list) and materiasAGuardar:
-            if materiasAGuardar and isinstance(materiasAGuardar[0], Materia):
-                materiasAGuardar = materiasADict(materiasAGuardar)
-            # Convertimos la lista de materias a un dict ordenado por clave y NRC
-            materiasAGuardar = clasificarMateriasEnDict(materiasAGuardar)
-        with open(RUTA_CATALOGO_MATERIAS, "w") as f:
-            f.write(json.dumps(materiasAGuardar, indent=4))
-
-    except Exception:
-        correcto = False
-        import traceback
-        print("Error: No se pudieron guardar las materias en el JSON")
-        print("Detalles:")
-        print(traceback.format_exc())
-
-    return correcto
-
-
-def materiasADict(materias: list, clasificarPorNrc: bool = False):
-    materias_dict = {} if clasificarPorNrc else []
-    for mat in materias:
-        if not isinstance(mat, Materia):
-            continue
-        mat_humana = mat.obtenerVersionHumana()
-        if clasificarPorNrc:
-            materias_dict[mat.nrc] = mat_humana
-        else:
-            materias_dict.append(mat_humana)
-    return materias_dict
-
-
-def dictAMaterias(materias: list, clasificarPorClave: bool = False):
-    materias_obj = {} if clasificarPorClave else []
-    for mat in materias:
-        req_keys = ['seccion', 'materia', 'cupo', 'disponible', 'creditos',
-                    'nrc', 'clave', 'sesion', 'horas', 'dias', 'edificio',
-                    'aula', 'periodo', 'profesor', 'url']
-        if not (isinstance(mat, dict) or req_keys in mat.keys()):
-            continue
-        horarios = []
-        # El valor por defecto no debería de obtenerse
-        centro = mat.get("centro", CENTRO_DEF)
-
-        for i in range(len(mat["sesion"])):
-            sesion = mat["sesion"][i]
-            horas = mat["horas"][i]
-            dias = mat["dias"][i]
-            edif = mat["edificio"][i]
-            aula = mat["aula"][i]
-            periodo = mat["periodo"][i]
-            horario = Horario(centro, sesion, horas, dias, edif, aula, periodo)
-            horarios.append(horario)
-
-        mat_obj = Materia(centro, mat["nrc"], mat["clave"],
-                              mat["materia"], mat["seccion"], mat["creditos"],
-                              mat["cupo"], mat["disponible"], horarios,
-                              mat["profesor"], mat.get("eliminada", False), mat["url"])
-        if clasificarPorClave:
-            if materias_obj.get(mat_obj.clave) is None:
-                materias_obj[mat_obj.clave] = {"nombre": mat_obj.obtenerNombre(), "grupos": {}}
-            materias_obj[mat_obj.clave]["grupos"][mat_obj.nrc] = mat_obj
-        else:
-            materias_obj.append(mat_obj)
-    return materias_obj
-
-
-def clasificarMateriasEnDict(listaMaterias: list):
-    materias_clasificadas = {}
-    for mat in listaMaterias:
-        if not materias_clasificadas.get(mat["clave"]):
-            materias_clasificadas[mat["clave"]] = {
-                "nombre": mat["materia"], "grupos": {}}
-        materias_clasificadas[mat["clave"]]["grupos"][mat["nrc"]] = mat
-    return materias_clasificadas
-
-
-def aplanarDictMaterias(dictMaterias: dict, retornarDict: bool = False):
-    materias_planas = {}
-    for materia in dictMaterias.values():
-        for grupo in materia["grupos"].values():
-            materias_planas[grupo["nrc"]] = grupo
-    return list(materias_planas.values()) if not retornarDict else materias_planas
-
-
-def obtenerMatsEliminadas(materiasAnteriores: dict, materiasActuales: dict):
-    materias_eliminadas = {}
-    # Iteramos sobre las materias actuales para ver cuáles ya no existen
-    for nrc, mat_ant in materiasAnteriores.items():
-        # Si no existe el NRC en las materias obtenidas de SIIAU, la borraron
-        if not nrc in materiasActuales.keys():
-            # Avisamos que la borraron
-            mat_ant.eliminada = True
-            # La añadimos a las que borraron
-            materias_eliminadas[nrc] = mat_ant
-
-    return materias_eliminadas
-
-
-def actualizarMaterias(materiasNuevas: dict, materiasAnteriores: dict):
-    materias_exportar = {}
-    # Iteramos sobre las materias nuevas para añadir/actualizar
-    for nrc, mat_nva in materiasNuevas.items():
-        # Verificamos si la materia nueva ya existía
-        # en la oferta para actualizarla
-
-        if nrc in materiasAnteriores.keys():
-            # Obtenemos la materia actual
-            mat_act = materiasAnteriores[nrc]
-            # Establecemos la url (o lo que tenía) la otra materia
-            mat_nva.url = mat_act.url
-            # Añadimos la materia
-            materias_exportar[nrc] = mat_nva
-
-        else:
-            # No existía, simplemente añadimos la materia
-            materias_exportar[nrc] = mat_nva
-
-    return materias_exportar
-
-
-def generarCatalogoMaterias(materiasNuevas: list, conservar_materias_eliminadas: bool = True):
-    # Hay catálogo de materias ya existente
-    if os.path.exists(RUTA_CATALOGO_MATERIAS):
-        # Indizamos las materias en un dict tipo {"nrc": materia...}
-        materias_actuales_obj = cargarMaterias(objetosMateria=True)
-        # Diccionario por comprensión
-        materias_nuevas_obj = {mat.nrc: mat for mat in materiasNuevas}
-        # Aquí se almacenarán las materias que vamos a exportar
-        materias_exportar = actualizarMaterias(
-            materias_nuevas_obj, materias_actuales_obj)
-
-        # Usualmente no queremos conservar las materias
-        # cuando ya empezó el ciclo y no las restablecieron
-        # De otro modo, puede que las restablezcan y hay que
-        # conservarlas
-        if conservar_materias_eliminadas:
-            materias_exportar.update(
-                obtenerMatsEliminadas(materias_actuales_obj,
-                                      materias_exportar))
-
-        # Convertimos el dict en lista para posterior procesamiento
-        materias = list(materias_exportar.values())
-
-    # No hay catálogo de materias y es el primero que generamos
-    else:
-        materias = materiasNuevas
-
-    # Guardamos el catálogo
-    guardarMaterias(materias)
 
 
 def extraerDatosFila(nodo: bs4.Tag):
@@ -480,7 +37,7 @@ def extraerDatosFila(nodo: bs4.Tag):
     return datos
 
 
-def extraerDatosTabla(tabla: bs4.Tag | bs4.NavigableString, saltarFilasInvalidas: bool = False, cantCeldasEsperadas: int = 0):
+def extraerDatosTabla(tabla: bs4.PageElement, saltarFilasInvalidas: bool = False, cantCeldasEsperadas: int = 0):
     tabla = tabla.find("tbody") if tabla.find("tbody") else tabla
     datosTabla = []
     for node in tabla.find_all("tr", recursive=False):
